Initial commit of mathymath
This is the initial import of mathymath for generating math
worksheets and quizes for elementary school math.  It consists of
a LaTeX file with a structured comment marking where the problems
go, a problem generator that generates random problems and produces
an output LaTeX file, and a "fortune" file to put a quote at
the bottom of each sheet.